chatbot04-app-ref/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
openai
google-generativeai
anthropic
requests
python-dotenv
marked